Fix emoji override
If two or more emoji font files are installed, the system perfers
the font that has longer sequence. However, if multiple emojis
are used in the single context, this fallback misinterpret the
continued emojis as sequence. Not to draw continued emojis with
unexpected fonts, split emoji with grapheme like rule.

Bug: 204329524
Test: minikin_tests
Test: atest CtsTextTestCases
Test: atest CtsGraphicsTestCases

Change-Id: Idaa2d7b86a43ae478760ba85dd23742b8b39f729