Introduce global MinikinFontFactory.
Before this CL, each Font instance had a TypefaceLoader pointer.
Because in reality we have only one TypefaceLoader (for MinikinFontSkia)
on Android, we can make it global.
Together with field reordering, this will save ~7KB (16 bytes * 431)
on oriole.

This CL introduces MinikinFontFactory, which bundles three functions
(read/skip/write) for serialization.
The caller of FontCollection::readVector and writeVector is responsible
for initializing the global MinikinFontFactory singleton.

We could remove this global MinikinFontFactory if we made minikin::Font
fully templated i.e. minikin::Font<class MinikinFontFactoryT>.
However, that would lead to a significant rework of Minikin library,
because Font is stored in global variables (SystemFont) and is depended
by many other classes (LayoutCache etc).

Bug: 174672300
Test: atest minikin_tests
Test: atest CtsGraphicsTestCases:android.graphics.cts.TypefaceTest
Change-Id: I1016358e4b618e0a1ac37bc6de07f93c326141b3