Remove mPos from BufferReader.
This CL replaces (mData, mPos) with mCurrent (== mData + mPos),
and fixes a minor issue with align() (that if mData was not aligned,
mData + align() was not aligned).

To that end, now BufferReader has two additional requirements:
(a) Alignment for each data record must be a power of 2 (2^n) and
    must be less than or equal to kMaxAlignment (== 8).
(b) The head address of the buffer must be aligned at kMaxAlignment.

This saves ~3KB on oriole (8 byte * 431 Font).

Bug: 174672300
Test: atest minikin_tests
Test: m -j
Change-Id: Ie070254e5e860ebe70e775b9ebe99c5e70b27e8e