Fix possible "new AxisTag[0]".
This is a follow up of
commit bbbdf1a6590064077466c2276f5c29278aaf16fb
to avoid calling "new AxisTag[0]".
libmemunreachable does not flag this case [1] but it is inefficient to
create a zero-length array as it calls malloc(0), which returns a
non-null value on Android.

[1] new T[0] and libmemunreachable:
First of all, "new T[0]" may or may not allocate a non-empty buffer
depending on type T. If T has a non-default destructor, "new T[0]"
allocates 8 bytes (on aarch64) to record the number of T elements
in the buffer. If T has a default destructor, "new T[0]" allocates 0
byte.

When T has a non-default destructor, new T[0] returns a pointer to the
end of the allocated buffer. That is,
  T* p = new T[0];
is equivalent to:
  void* buf = malloc(8);
  *((size_t*)buf) = 0;
  T* p = (T*)(((intptr_t)buf) + 8);
In this case, libmemunreachable flags 'buf' as unreachable even if
'p' is reachable, because it does not "consider a pointer to the end of
a buffer to be inside the buffer" [2].

On the other hand, if T has a default destructor,
  T* p = new T[0];
is equivalent to:
  void* buf = malloc(0);
  T* p = (T*)buf;
In this case, libmemunreachable considers 'buf' as reachable from 'p',
because libmemunreachable internally treats buf + 1 as the end of
the buffer [3].

[2] https://cs.android.com/android/platform/superproject/+/master:system/memory/libmemunreachable/HeapWalker.h;drc=8837c7216fb146ba80cdcf14575483aa45735f88;l=125

[3] https://cs.android.com/android/platform/superproject/+/master:system/memory/libmemunreachable/HeapWalker.cpp;drc=85ad401b7738905aeb2a4ebbe35045f9703ef4dd;l=47

Bug: 174672300
Test: atest minikin_tests
Change-Id: Icc971429f6cd6e975a8a71fd2e71ade37c4f0612