Do not distibute the width to the control characters
Roboto font will have a zero-width glyph for the control characters
for avoiding tofu rendering. This will confuse the minikin width
distribution logic used for placing cursor in the middle of the
ligature.

The control characters should not have widths, so just excluding
the control characters from the width distribution logic.

Bug: 241503271
Test: Manually done
Change-Id: Ibce61016fca7a8fe075e266e111927c43b5c4729