Add default family flag to the font family
The default family indicates that the default font to be
used for named font families. The named font family is used as a
highest priority family, like developer provided font family.
Before this change, it is determined if the family is located at
the top of the fallback list, it is treated as a default font.
Now, the named font family can be mulitple font families and they
are all needed to be treated as the highest priority font famlies.
To identify the families in the default family list.

Bug: 249787583
Test: atest TypefaceSystemFallbackTest FontListParserTest
Test: atest UpdatableFontDirTest UpdatableSystemFontTest
Test: atest GtsFontHostTestCases FontManagerTest
Test: minikin_tests
Change-Id: If01378628a0ed3f2e02bc0f346f2d05e76bd77bc